style: create separate pages for sections
