+++ src/App.jsx
@@ -1,18 +1,25 @@
-import './App.css'
+import React from "react";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Navbar from "./components/Navbar/Navbar.jsx";
-import Hero from "./components/Hero/Hero.jsx";
-import Projects from "./components/Projects/Projects.jsx";
-import Skills from "./components/Skills/Skills.jsx";
-import Footer from "./components/Footer/Footer.jsx"
+import Home from "../pages/Home/Home.jsx";
+import Projects from "../pages/Projects/Projects.jsx";
+import Skills from "../pages/Skills/Skills.jsx";
+import Footer from "./components/Footer/Footer.jsx";
 
 function App() {
-  return<>
-    <Navbar/>
-    <Hero/>
-    <Projects/>
-    <Skills/>
-    <Footer/>
-  </>
+  return (
+      <Router>
+        <Navbar />
+
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/projects" element={<Projects />} />
+          <Route path="/skills" element={<Skills />} />
+        </Routes>
+
+        <Footer />
+      </Router>
+  );
 }
 
-export default App
+export default App;
